refact/game_teams class added, necessary methods moved
