+++ lib/reusable.rb
@@ -72,4 +72,24 @@ module Reusable
     opponents_games_hash = opponent_games_arr.tally
     opponents_games_hash
   end
+
+  def games_by_season(season)
+    games_in_season = @games.games.collect { |row| row[:game_id] if row[:season] == season}
+    games_in_season.compact
+  end
+
+  def team_average_number_of_goals_per_game(team_id)#Helper--game_teams
+    @game_count = 0
+    @game_score = 0
+    @game_teams.each do |row|
+      if row[:team_id].to_i == team_id.to_i
+        @game_count += 1
+        @game_score += row[:goals].to_i
+      end
+    end
+    if @game_count > 0
+      return @game_score.to_f / @game_count.to_f
+    end
+  end
+
 end
