+++ lib/stat_tracker.rb
@@ -1,6 +1,6 @@
 require 'CSV'
 require_relative './games'
-# require './lib/league'
+require_relative './game_teams'
 require_relative './teams'
 require_relative './reusable'
 class StatTracker
@@ -9,10 +9,10 @@ class StatTracker
 
   def initialize(data1, data2, data3)
     @team_csv = data2
-    @game_teams = data3
     @teams_hash = team_hash
     @games = Games.new(data1, @teams_hash)
     @teams = Teams.new(data2, @teams_hash)
+    @game_teams = GameTeams.new(data3, @teams_hash, @games)
   end
 
   def self.from_csv(locations)
@@ -44,66 +44,16 @@ class StatTracker
     @games.percentage_ties
   end
 
-  def games_by_season(season)
-    games_in_season = @games.games.collect { |row| row[:game_id] if row[:season] == season}
-    games_in_season.compact
-  end
-
   def most_accurate_team(season_id)
-    hash = {}
-    final_hash = {}
-    season_games = games_by_season(season_id)
-    array = @game_teams.select { |row| season_games.include?(row[:game_id])}
-    array.each do |row|
-      if !hash.keys.include?(row[:team_id])
-        hash[row[:team_id]] = {goals: row[:goals].to_f, shots: row[:shots].to_f}
-      else
-        hash[row[:team_id]][:goals] += row[:goals].to_f
-        hash[row[:team_id]][:shots] += row[:shots].to_f
-      end
-    end
-    hash.each do |key, hash2|
-      final_hash.merge!("#{key}" => (hash2[:goals] / hash2[:shots]))
-    end
-    final_hash.each do |k, v|
-      if v == final_hash.values.max
-        return @teams_hash[k]
-      end
-    end
+    @game_teams.most_accurate_team(season_id)
   end
 
   def least_accurate_team(season_id)
-    hash = {}
-    final_hash = {}
-    season_games = games_by_season(season_id)
-    array = @game_teams.select { |row| season_games.include?(row[:game_id])}
-    array.each do |row|
-      if !hash.keys.include?(row[:team_id])
-        hash[row[:team_id]] = {goals: row[:goals].to_f, shots: row[:shots].to_f}
-      else
-        hash[row[:team_id]][:goals] += row[:goals].to_f
-        hash[row[:team_id]][:shots] += row[:shots].to_f
-      end
-    end
-    hash.each do |key, hash2|
-      final_hash.merge!("#{key}" => (hash2[:goals] / hash2[:shots]))
-    end
-    final_hash.each do |k, v|
-      if v == final_hash.values.min
-        return @teams_hash[k]
-      end
-    end
+    @game_teams.least_accurate_team(season_id)
   end
 
   def average_win_percentage(team_id)
-    counter = 0
-    team_id_h = @game_teams.group_by { |row| row[:team_id].itself }
-    team_id_h[team_id].each do |row|
-      if row[:result] == "WIN"
-        counter += 1
-      end
-    end
-    (counter.to_f / team_id_h[team_id].count.to_f).round(2)
+    @game_teams.average_win_percentage(team_id)
   end
   #C O O O O O O O O O O O O LIN
   def average_goals_per_game
@@ -112,7 +62,7 @@ class StatTracker
 
   def average_goals_by_season
     average_by_season = {}
-    season_hash = @games.games.group_by { |row| row[:season].itself }
+    season_hash = seasons_hash
     season_hash.each do |season, games|
       counter = 0
       game = 0
@@ -142,7 +92,7 @@ class StatTracker
   game_array = []
   tackle_hash = {}
   max_tackles_team = []
-  @game_teams.each do | row |
+  @game_teams.game_teams.each do | row |
       if season.to_s.include?(row[:game_id][0..3])
       game_array << row
       end
@@ -173,7 +123,7 @@ class StatTracker
   working_array = seasons_hash
   acceptable_games = []
   (working_array[season]).each { |row| acceptable_games << row[:game_id]}
-  @game_teams.each do | row |
+  @game_teams.game_teams.each do | row |
       if acceptable_games.include?(row[:game_id])
       game_array << row
       end
@@ -202,75 +152,19 @@ class StatTracker
   end
 
   def most_goals_scored(team_id)
-    goals = []
-    most_goals = @game_teams.group_by { |row| row[:team_id].itself}
-    most_goals.each do | team, stats |
-      if team_id.to_s == team
-        stats.each do | goal |
-          goals << goal[:goals].to_i
-        end
-      end
-    end
-    return goals.sort.pop
+    @game_teams.most_goals_scored(team_id)
   end
 
   def fewest_goals_scored(team_id)
-    goals = []
-    fewest_goals = @game_teams.group_by { |row| row[:team_id].itself}
-    fewest_goals.each do | team, stats |
-      if team_id.to_s == team
-        stats.each do | goal |
-          goals << goal[:goals].to_i
-        end
-      end
-    end
-    return goals.sort.shift
+    @game_teams.fewest_goals_scored(team_id)
   end
   # T H I A G O O O O O O O A L L L L L
   def winningest_coach(season_id)
-    hash = {}
-    final_hash = {}
-    season_games = games_by_season(season_id)
-    array = @game_teams.select { |row| season_games.include?(row[:game_id])}
-    array.each do |row|
-      if !hash.keys.include?(row[:head_coach])
-        hash[row[:head_coach]] = [row[:result]]
-      else
-        hash[row[:head_coach]] << row[:result]
-      end
-    end
-    hash.each do |key, array|
-      wins = array.select { |element| element == "WIN" }
-      final_hash.merge!("#{key}" => (wins.count.to_f / array.count.to_f))
-    end
-    final_hash.each do |k, v|
-      if v == final_hash.values.max
-        return k
-      end
-    end
+    @game_teams.winningest_coach(season_id)
   end
 
   def worst_coach(season_id)
-    hash = {}
-    final_hash = {}
-    season_games = games_by_season(season_id)
-    array = @game_teams.select { |row| season_games.include?(row[:game_id])}
-    array.each do |row|
-      if !hash.keys.include?(row[:head_coach])
-        hash[row[:head_coach]] = [row[:result]]
-      else
-        hash[row[:head_coach]] << row[:result]
-      end
-    end
-    hash.each do |key, array|
-      win = array.select { |element| element == "WIN" }
-      final_hash.merge!("#{key}" => (win.count.to_f / array.count.to_f))
-    end
-    final_hash.each do |k, v|
-      if v == final_hash.values.min
-        return k
-      end
-    end
+    @game_teams.worst_coach(season_id)
   end
 
   def highest_scoring_visitor
@@ -311,35 +205,13 @@ class StatTracker
     @teams.count_of_teams
   end
 
-  def team_average_number_of_goals_per_game(team_id)#Helper--game_teams
-    @game_count = 0
-    @game_score = 0
-    @game_teams.each do |row|
-      if row[:team_id].to_i == team_id.to_i
-        @game_count += 1
-        @game_score += row[:goals].to_i
-      end
-    end
-    if @game_count > 0
-      return @game_score.to_f / @game_count.to_f
-    end
-  end
 
   def best_offense
-    @id_avg_hash = {}
-    @teams_hash.each do |k,v|
-      @id_avg_hash.merge!("#{@teams_hash[k]}" => team_average_number_of_goals_per_game(k))
-    end
-    hash_max(@id_avg_hash)
-
+  @game_teams.best_offense
   end
 
   def worst_offense
-    @id_avg_hash = {}
-    @teams_hash.each do |k,v|
-      @id_avg_hash.merge!("#{@teams_hash[k]}" => team_average_number_of_goals_per_game(k))
-    end
-    hash_min(@id_avg_hash)
+    @game_teams.worst_offense
   end
 
   def team_info(team_id)
